rename hseries.rs to qsystem.rs
