+++ tket2-hseries/README.md
@@ -3,7 +3,7 @@
 ![msrv][]
 
 A TKET2 tool for preparing and validating `Hugr`s for compilation targeting
-Quantinuum H-series  quantum computers.
+Quantinuum System  quantum computers.
 
 ## Usage
