+++ tket2-hseries/src/bin/tket2-hseries.rs
@@ -10,7 +10,7 @@ fn main() {
             let reg = ExtensionRegistry::try_new([
                 tket2::extension::TKET2_EXTENSION.to_owned(),
                 tket2::extension::rotation::ROTATION_EXTENSION.to_owned(),
-                tket2_hseries::extension::hseries::EXTENSION.to_owned(),
+                tket2_hseries::extension::qsystem::EXTENSION.to_owned(),
                 tket2_hseries::extension::futures::EXTENSION.to_owned(),
                 tket2_hseries::extension::result::EXTENSION.to_owned(),
             ])
