+++ tket2-hseries/src/lazify_measure.rs
@@ -27,7 +27,7 @@ use lazy_static::lazy_static;
 
 use crate::extension::{
     futures::FutureOpBuilder,
-    hseries::{self, HSeriesOpBuilder},
+    qsystem::{self, HSeriesOpBuilder},
 };
 
 /// A `Hugr -> Hugr` pass that replaces [tket2::Tk2Op::Measure] nodes with
@@ -114,7 +114,7 @@ lazy_static! {
         let [qb, lazy_r] = builder.add_lazy_measure(qb).unwrap();
         let [r] = builder.add_read(lazy_r, BOOL_T).unwrap();
         builder
-            .finish_hugr_with_outputs([qb, r], &hseries::REGISTRY)
+            .finish_hugr_with_outputs([qb, r], &qsystem::REGISTRY)
             .unwrap()
     };
 }
@@ -146,7 +146,7 @@ fn measure_replacement(num_dups: usize) -> Hugr {
     assert_eq!(num_out_types, rs.len());
     assert_eq!(num_out_types, num_dups + 1);
     builder
-        .finish_hugr_with_outputs(rs, &hseries::REGISTRY)
+        .finish_hugr_with_outputs(rs, &qsystem::REGISTRY)
         .unwrap()
 }
 
@@ -223,14 +223,14 @@ mod test {
 
     use crate::extension::{
         futures::{self, FutureOpDef},
-        hseries::HSeriesOp,
+        qsystem::HSeriesOp,
     };
 
     use super::*;
 
     lazy_static! {
         pub static ref REGISTRY: ExtensionRegistry = ExtensionRegistry::try_new([
-            hseries::EXTENSION.to_owned(),
+            qsystem::EXTENSION.to_owned(),
             futures::EXTENSION.to_owned(),
             TKET2_EXTENSION.to_owned(),
             PRELUDE.to_owned(),
